[DOC][REFACTORING] Renamed r-lint to rmageddon
