--- r_lint/__init__.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python
-""" Main r_lint package file """
-
-import pkg_resources
-
-__version__ = pkg_resources.get_distribution("r_lint").version
