+++ rmageddon/__init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+""" Main rmageddon package file """
+
+import pkg_resources
+
+__version__ = pkg_resources.get_distribution("rmageddon").version
