+++ setup.py
@@ -15,20 +15,20 @@ with open('requirements.txt') as f:
     required = f.read().splitlines()
 
 setup(
-    name = 'r-lint',
-    version = version,
-    description = 'Small linting tool for R containers at QBiC',
-    long_description = readme,
-    keywords = ['R', 'linting', 'lint', 'Docker', 'container'],
-    author = 'Sven Fillinger',
-    author_email = 'sven.fillinger@qbic.uni-tuebingen.de',
-    license = license,
-    scripts = ['scripts/r-lint'],
-    install_requires = required,
+    name='rmageddon',
+    version=version,
+    description='Small linting and building tool for R containers at QBiC',
+    long_description=readme,
+    keywords=['R', 'linting', 'lint', 'Docker', 'container'],
+    author='Sven Fillinger',
+    author_email='sven.fillinger@qbic.uni-tuebingen.de',
+    license=license,
+    scripts=['scripts/rmageddon'],
+    install_requires=required,
     setup_requires=[
-        'twine>=1.11.0',
-        'setuptools>=38.6.',
-    ] + ([] if sys.version_info.minor == 4 else ['wheel>=0.31.0']),
-    packages = find_packages(exclude=('docs')),
-    include_package_data = True
+                       'twine>=1.11.0',
+                       'setuptools>=38.6.',
+                   ] + ([] if sys.version_info.minor == 4 else ['wheel>=0.31.0']),
+    packages=find_packages(exclude='docs'),
+    include_package_data=True
 )
